Added StandAlone folder for some autonomous scripts.
Implemented the LabJack acquisiton card, as a sensor and also as an Actuator.

Modified the measureComediByStep block to work also with LabJack devices, and renamed it measureByStep.
The old measureComediByStep is still there (and works)  but should not be used anymore to simplify.

the Streamer blocks as been modified to accept LabJack devices.

Crappy is now in version 1.1.4