[ref] Removing NotInstalled class
In order to build the documentation with Sphinx (coming soon), all the
classes must be defined. This is why we are getting rid of the
NotInstalled placeholder for classes that required optional modules.
Instead, the classes will all be imported and will raise an error when
instanciated if the optional dependencies are not met. This is done by
replacing the optional module by an instance if the class OptionalModule
(see _global.py). This class will raise an error with a convenient
message when celling __getitem__ instead, handling the optional
dependencies at runtime.